Add blocks and depends_on fields to the Review Request resource.
The new support for the Depends On and Blocks fields didn't add any API
for getting these values to the Review Request resource. They've now
been added. Along with this, the documentation for the depends_on field
in the Review Request Draft resource has been fixed to point to the
right type.