Move the reply draft banner out into its own view.
The reply draft banner has been moved out of reviews.js and into
ReviewReplyDraftBannerView. The new code is more "dumb" in that it
doesn't deal with anything beyond rendering and handling the
publish/discard. It's up to the caller to determine what happens after
(such as page reloading).

The floating support moved into FloatingBannerView, which
ReviewReplyDraftBannerView inherits from. This has been made a bit more
generic, so that it can be use for other banners later on.