Don't reload the page when publishing/discarding a reply.
Publishing/discarding a reply to a review no longer reloads the review
request page. That was ugly and made it easy to lose your place on the
page, especially for large review requests.

Now, discarding just removes all the comments that were discarded, and
publishing turns all the draft comment entries into published comment
entries.

As part of this, I've also cleaned up a lot of the templating around
these comments. We used to have hacks to look for IDs starting with
"yourcomment", which was stupid. We also had other IDs sticking around.
None of those are needed anymore, nor are some classes, so they've all
been removed.