Rewrite the review dialog.
The review dialog has had a full rewrite. It's no longer just a
container for a server-side rendered form, which was done in a very
hacky way and could be slow to load for large changes. Now it displays
instantly and fetches the comments itself, rendering them one at a time.

This looks the same as the old form, but with one improvement:
File attachment thumbnails are now rendered as well. They appear just
like screenshot thumbnails do. Previously, we'd only render the
caption or filename.

Unit tests were added for this.