Make expandedFields a BaseResource concept.
DraftReviewRequest had a concept of expanding fields in a request by
populating expandedFields. This turns out to be useful elsewhere, so
move this logic out and into BaseResource.

Now, any resource class can set the list of fields to expand, in order
to provide more detailed information or child resources to the callers.