Fix a logic error for inserts in the new improved interdiff filter.
The new interdiff filter had problems dealing inserts in the right-hand
side of an interdiff. The way it was computing whether to switch to the
next range was incorrect. It was considering inserts as equals.

Now it matches the logic on the left-hand side.

A unit test was added based on real interdiff opcodes for a change where
this manifested.

I also noticed that my original test case did in fact have its one and
only right-hand-side insert missing, which was restored after this
change.