Don't set options.data when using a form, and fix tests.
The new change to update options.data.expand and ensure options.data is
a dictionary shouldn't apply to forms. We now only set this if it's not
a form.

Also, a couple unit tests were still thinking it should be undefined.
It's fixed now to check instead of it's an empty object.