Improve the style of the issue summary table.
This makes a few tweaks to the issue summary table. There's now more
padding in each row, making things less cramped. The double border has
also been removed, cleaning up the look. The "resolved" border is also a
lot less faint, blending in better with the other rows.

The description has been truncated at 20 words, which seems to be a
decent length. This keeps it from being a squashed mess.

The text in the cells are aligned to the top now, which helps with
wrapping. The description column now takes all the available space.

The Issue ID column has been removed. It wasn't really that useful and
didn't cleanly map (since the concept of an issue ID didn't exist
anywhere else).