Move from reviewableIDField to defaultCommentBlockFields in Reviewables.
reviewableIDField was used to indicate which ID field in the
Reviewable's attribute list should be populated into any newly created
CommentBlocks by default. This wasn't very clearly named, though, and
wasn't sufficient for diffs. With diffs, you have two identifying pieces
of information: The filediff's ID, and possibly the interfilediff's ID.

This new field, defaultCommentBlockFields, is a list (or optionally a
function returning a list), and can be used for more than the ID, if
needed.

This has backwards-compatibility with existing reviewables, but will log
a deprecation warning.