Move diff reviewable setup and navigation/anchor selection out of diffviewer.js.
DiffViewerPage is now responsible for all keyboard navigation around the
diffs. It makes use of the new KeyBindingsMixin. It also owns anchor
selection, which (currently) both diffviewer.js and DiffReviewableView
make use of through a temporary RB.scrollToAnchor function.

DiffViewerPage also now sets up the DiffReviewableView for a loaded diff
fragment. This was tied closely enough to the anchor setup that it
seemed worth including in the same change.