Add an API resource to verify diff files.
This change adds a resource to /api/verify-diff/ which accepts POST requests
with arguments similar to DiffResource, but does not actually create a diff.
This allows prospective clients (such as the web UI or rbtools) to validate
whether a diff file will be usable before creating a review request for it.

Testing done: Ran unit tests.

Reviewed at http://reviews.reviewboard.org/r/4252/