Fix unwanted saves when setting/getting ReviewRequest.commit.
ReviewRequest.commit behaved unlike any other attribute of a model.
Getting or setting it would trigger a save of the model.

This caused two problems.

First, we could end up loading a ReviewRequest and getting the commit,
triggering a save too early, and messing up state. The webapi docs
generation was impacted by this, and it led to failures.

Second, call sites that updated the commit and other properties resulted
in two database hits.

We no longer automatically save when setting. We still do migrate the
changenum to commit_id when getting, if needed, but we do it as an
atomic change in the database and not a full model save.