Fix broken/missing docs and behavior for the new validation API.
The new validation API was missing documentation. While adding it, I
noticed a few problems with how the API was being represented.

The resource name "diffs" collided with the other "diffs" resource, and
overwrote the URI template in the root resource. Resource names should
be unique, and more specific, so I renamed this to "diff_validation." I
also got rid of the mimetype, letting it build upon the new name (and
fixing some oddity with the resource not properly encoding in XML, for
some reason).

I also noticed any client following the link from /validation/ to
/validation/diffs/ would get an error, since there's no GET handler. A
proper REST resource should provide some representation. So, we provide
a payload containing the typical links for getting/posting to the
resource.

I also fixed up some documentation issues, some doc generation
issues with singletons, and added the pages for the validation resource.

Note that this does change the payload key for the diff validation
response.