Provide ReviewRequests and Reviews to reviewables and comment blocks.
Reviewables and CommentBlocks now take 'reviewRequest' and 'review'
attributes, instead of having to call out to gReviewRequest.

The 'reviewRequest' attribute is required, but 'review' is not. If not
provided, AbstractReviewable will get/create a draft Review object from
the ReviewRequest. These two parameters are now required.

Reviewables also now have a convenience function for creating
CommentBlock and adding them. This will do the work of passing along the
ReviewRequest and Review to the CommentBlock instance.

Existing Reviewables and CommentBlocks have been updated to work with
these changes. External ones will need to be changed.