Add repositories/<id>/branches/ and repositories/<id>/commits resources.
This change adds APIs for getting a list of all branches on a repository, and a
(paginated) list of commits backwards in time from a given starting point. These
are currently implemented for the SVN SCMTool and the GitHub hosting service.

Testing done:
- Ran unit tests.
- Used these in conjunction with my post-commit implementation.

Reviewed at http://reviews.reviewboard.org/r/4265/