Fix an issue with validating against repos that don't use basedir.
The UploadDiffForm had a buried case where it set the basedir to the empty
string in cases where the repository uses absolute paths, and it turns out that
this is necessary because we pass the value to os.path.join(), which really
doesn't like getting None for its first argument. I had missed this when
implementing ValidateDiffResource, which meant it worked for svn but didn't for
git.

Testing done: validated some diffs against a git repository.
Reviewed at http://reviews.reviewboard.org/r/4268/