Ensure the local_id update is committed to the server.
When we create a ReviewRequest on a server using a LocalSite, we set a
local_id based on previous IDs. This has worked in the past, but
something subtle broke it. We need to ensure this change is committed to
the server before any further use. We now commit the transaction and
then continue. This will prevent ReviewRequests that have a LocalSite
without a Local ID set from showing up in the database.