Add a $.fn.bindClass function for mapping an attribute to a class.
bindClass is used when you want to toggle whether a class is on an
element based on the value of a boolean attribute. It adds or removes
the class initially based on the attribute at the time of binding, and
responds to all updates to the property.

Like the other binding functions, it also supports an 'inverse' option
for flipping when the class will be present.