Move the diff viewer whitespace toggles into the new objects.
The Show Whitespace Changes and Show Extra Whitespace toggles now live
in the new objects and interact with the DiffReviewableViews to change
the state.

The cookie management now lives in UserSession. Callers just need to set
an attribute and it will update the cookie. This change introduces
jquery.cookie for actually setting and getting cookies, because
JavaScript's native cookie support is terrible.