Fix "whitespace-only changes" toggle for files with only whitespace changes.
A change made a while back broke the logic for determining when to show
the "This file has only whitespace changes" description. That
description was necessary, because we keyed off the visibility toggles
for whitespace-only chunks based on it. So, the entire feature was
broken.

That description element has been restored and the toggle works again.

Fixes bug #2941