Revert my prior keybinding fixes and fix the root cause of problems.
The previous fixes weren't fixing the root cause of the problem, which
was that the bindings would end up removed prematurely, due to the order
in which the delegate/undelegate functions were called. I've reverted
the previous fixes and gone back to listening on this.$el, and fixed the
order of delegate/undelegate.