Adjust the way that parent revision numbers are computed for SVN.
In the get_change implementation for SVN, I was just subtracting one from the
revision number to get the parent to diff against. This works for a typical SVN
server, but I was worried that it might fall down for wacky things like GitHub's
subversion layer. I've changed it to use the cached parent if available, and if
not, use the next item in the log.

Testing done: Ran unit tests.
Reviewed at http://reviews.reviewboard.org/r/4316/