Add a field to DiffSet for storing the commit ID a change is based on.
This field can be set when uploading a diff via the API. We can make use
of it for certain hosting services that only provide file fetching APIs
that take Git commit SHA1s instead of blob SHA1s.

The field is optional. RBTools will be updated to set it for Git, but it
will otherwise be used only as needed.