Add JavaScript for validating diff files.
This is the client side of the diff validation resource. To use this, one
instantiates the model, sets the repository/diff/parent diff/basedir, and calls
save(). Either the "success" or "error" callbacks in the options object passed
to save will be called, based on whether the diff parses correctly and the
source files exist.

Testing done:
- Ran unit tests.
- Used this with a number of different diffs that caused various errors.

Reviewed at https://reviews.reviewboard.org/r/4320/