Add a commit selector view.
This is used by the new new review request UI for creating requests from
committed revisions. This handles the basic functionality of displaying a list
of commits and triggering a 'create' event when one is clicked (or navigating to
the review request if one already exists).

There's a little bit more on my todo list for this, namely:
- Providing some visual indication when there's already a review request.
- Using the API pagination to provide infinite scroll.

Testing done: Used this in my post-commit UI.

Reviewed at https://reviews.reviewboard.org/r/4322/