Escape text on hidden content dividers.
ReviewBoard shows you a certain amount of context around the changes.
Above
this you're provided with two options...

  [+] 1005 lines
  [+] public void titlesModified(List<Title> titlesAdd, List<String>
      titlesDelete,

This is all fine and well, except in the case of content like that shown
above.  'Title' coincidently belongs to an html tag, and including it
breaks the html parser for the page (as it would with '</script>' and a
few other things). HTML escaping these lines to prevent this.

The main visible result is that diff hunks beyond the first fail to
load. In firebug you get the error of...

Uncaught ReferenceError: setFileAttachmentComments is not defined

Reviewed at http://reviews.reviewboard.org/r/4330/
Fixes bug #2612