Update Amazon S3 support to use the newer django-storages s3boto backend.
It seems that the Amazon S3 support for file storage is broken. It is
likely that this is due to the changes in Amazon side. Right now the
default AWS SDK for Python is boto which is already supported by
django-storages. it only requires to specify a different backend.

This will add a new dependency to reviewboard in order to support
Amazon S3. Apart from django-storages, boto package is also required.

Note: I needed to uninstall the botocore package which was also
installed in my system. It seems there was a conflict with boto.

Reviewed at http://reviews.reviewboard.org/r/4333/