Combine all our icons into a spritesheet.
This spritesheet contains a good chunk of the artwork Review Board makes
use of. It speeds up load and render time, since we get to replace 26
file downloads with 1.

Some images were thrown out, such as the icons for certain widgets in
the admin UI that weren't really necessary, and the dropdown triangle on
the action menus.

In a few cases, logic or style changes had to be made to work with the
new class names and icon divs. The only cases where things differ is
where we had some pretty bad alignment choices from way back when on a
couple elements, and they're now fixed. (Pencil on the summary being one
example.)

This work will make it easier for us to add retina icons in time.