Add retina versions of all our icons.
I've redone all the icons as SVGs and fixed up parts of the icons that I
thought could be improved. With this SVG master, we can now generate 2x
icons for Retina displays.

The .rb-icon rule has been updated to select icons from the new
spritesheet when on a Retina display.