Upgrade Flot to 0.8.1.
We haven't updated Flot in a while. The newer versions support Retina
displays, which really improves the look of graphs in the admin UI.

None of our code has to change, though we do need to add the flot time
plugin for the Review Board Activity widget, as this is no longer part
of Flot.