Allow commenters to change the status of the issues they open.
Until now, only the owner of a review request could change the status of an
issue. A frequently recurring request is to allow the person who originally
opened the comment to do the same--especially to be able to reopen it.

This is an update of Niklas Hambuechen's change at
https://reviews.reviewboard.org/r/3837/

I've updated it to address the original concerns with that change, as well as
streamlined a lot of the code.

Fixes bugs 2713 and 2888.

Testing done:
- Ran unit tests.
- Verified that the "interactive" parameter in the template was correct.

Reviewed at http://reviews.reviewboard.org/r/4380/