Move the CommentBarIssueView creation code into ReviewBoxView.
The code to create a CommentBarIssueView was being done inside the
template for the page, adding more script tags and more document.ready
calls.

This code is now created when rendering the ReviewBoxView. It creates
based on data attributes now stored on the element. It's part of an
effort to remove <script> tags in the templates, which will help with
the require.js migration.