Fix default value for HostingService._http_request headers parameter
The default value for the 'headers' parameter to the _http_request was None,
which meant that if something called it without passing in headers,
urllib2.Request()'s constructor would throw an exception. Other similar code
uses a default value of an empty dictionary, which is more appropriate.

Reviewed at https://reviews.reviewboard.org/r/4392/