Allow FileAttachment.file to store longer file paths.
FileField, by default, limits to 100 characters, which was barely enough
as it was. With the new change for prepending a hash, this was
suffocating.

We now set that max length to 512 characters. While that's far smaller
than the max path length of most file systems, it should be enough for
our use, and is far more than what users have had in the past.