Fix ValueError when the user enters a bad cache host.
If the user enters a cache hostname which doesn't split properly, it makes the
admin site unavailable. This is because the cache stats code was just blindly
trying to unpack the values without any error handling. I've updated it to
cache this case, log, and continue.

Fixes bug 3049.
Reviewed at https://reviews.reviewboard.org/r/4402/