Pass on base_commit_id to hosting services for file lookups.
The base_commit_id is now passed on to the get_file/get_file_exists
functions in hosting services, so that they may make use of it when
fetching file data.

It's not passed on to SCMTools, since we currently have no need there,
but it can be added later if needed.

This, combined with some upcoming changes, will allow us to support
services like Beanstalk and BitBucket for Git.