Add infinite scroll for commits in the new review request UI.
This change makes it so scrolling to within 50px of the bottom of the list of
commits in the New Review Request page will automatically load another page.
This shows the little "Loading..." spinner at the top, but is otherwise pretty
seamless.

Testing done:
- Loaded many pages of commits.
- Connected a repo with only a small number of commits, and saw that it stopped
  loading once I hit the bottom.

Reviewed at https://reviews.reviewboard.org/r/4411/